% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDepth_asymp.R
\name{calcDepth_asymp}
\alias{calcDepth_asymp}
\title{Calculating the K-sign depth (asymptotically)}
\usage{
calcDepth_asymp(res, K, transform = FALSE, exact = TRUE)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{K}{[\code{integer(1)}]\cr
the parameter K for the sign depth}

\item{transform}{[\code{logical(1)}]\cr
Should the depth be transformed by the formula \eqn{N * (res - 0.5^(K-1))},
so that the distribution of the K-sign depth converges for large number
of data points N? Default is \code{FALSE}.}

\item{exact}{[\code{logical(1)}]\cr
Should the depth be calculated exactly? If \code{FALSE}, terms of scale
o(1) will ne ignored. At the moment, for exact
results this method is available only for K = 2, 3, 4, 5. Default is
\code{TRUE}.}
}
\value{
[\code{numeric(1)}] the calculated K-sign depth of \code{res}.
}
\description{
\code{calcDepth_asymp} calculates the K-sign depth of a given
vector of residuals (asymptotically).
}
\note{
The implementation is based on a work of Dennis Malcherczyk.
}
\examples{
calcDepth_asymp(rnorm(10), 3)
calcDepth_asymp(runif(100, -1, 1), 4, transform = TRUE)
calcDepth_asymp(runif(100, -1, 1), 4, transform = TRUE, exact = FALSE)
}
\references{
Malcherczyk, D. (2021+). K-sign depth: Asymptotic distribution, efficient
computation and applications. Dissertation in preparation.
}
