% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcSimpDepth}
\alias{calcSimpDepth}
\title{Simplified K-sign depth}
\usage{
calcSimpDepth(res, K)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{K}{[\code{integer(1)}]\cr
the parameter K of the simplified K-sign depth}
}
\value{
[\code{numeric(1)}] the calculated simplified K-sign depth of res.
}
\description{
\code{calcSimpDepth} calculates the simplified K-sign depth of a given vector of residuals.
}
\examples{
calcSimpDepth(rnorm(10), 3)
}
