% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthtest.R
\name{depth.test}
\alias{depth.test}
\alias{depth.test.default}
\alias{depth.test.formula}
\title{Sign Depth Test}
\usage{
depth.test(x, ...)

\method{depth.test}{default}(x, k, ...)

\method{depth.test}{formula}(formula, data, params, k, ...)
}
\arguments{
\item{x}{[\code{numeric}]\cr vector of residuals.}

\item{...}{[any] further arguments in the formula-method. Arguments are passed
to the \code{multiSorting}-function.}

\item{k}{[\code{integer(1)}]\cr parameter of the Sign Depth.}

\item{formula}{[\code{formula}]\cr a formula of the form \code{lhs ~ rhs} where
lhs is a numeric giving the target variable and rhs are the influencing variables .}

\item{data}{[\code{data.frame}]\cr a data frame containing the variables in the formula.}

\item{params}{[\code{numeric}]\cr a parameter vector for the null-hypotheses.}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
 \item{\code{statistic}}{the value of the depth.}
 \item{\code{parameter}}{the k of the Sign Depth.}
 \item{\code{p.value}}{the p-value for the test.}
 \item{\code{alternative}}{a character string describing the alternative hypothesis.}
 \item{\code{data.name}}{a character string giving the name of the data.}
 \item{\code{method}}{a character string describing the method.}
}
}
\description{
Performs a robust test whether a statistical model fits to some data.
}
\details{
In the default method a residual vector is given and it is tested whether this
residual vector comes from a model that describes the data sufficiently.
With the formula interface it is possible to give a formula and a parameter vector,
so that a linear model with the given parameters is fitted and the residuals are
calculated with this model.

At the moment only linear models for the formula interface are supported!

The parameter k of the Sign Depth has to be a natural number. Currently
only \eqn{k = 2, 3, 4, 5} is supported!

The quantiles used for calculating the p-value of the test are simulation based.
For more information see \code{\link{qdepth}}.
}
\examples{
depth.test(rnorm(30), 3)
depth.test(y ~ ., data = data.frame(x = rnorm(30), y = rnorm(30)),
  params = c(1, 1), k = 3)

}
\seealso{
\code{\link{qdepth}}, \code{\link{calcDepth}}, \code{\link{multiSorting}},
\code{\link{sign.test}}, \code{\link{f.test}}
}
