% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest.R
\name{f.test}
\alias{f.test}
\title{F Test}
\usage{
f.test(formula, data, params)
}
\arguments{
\item{formula}{[\code{formula}]\cr a formula of the form \code{lhs ~ rhs} where
lhs is a numeric giving the target variable and rhs are the influencing variables .}

\item{data}{[\code{data.frame}]\cr a data frame containing the variables in the formula.}

\item{params}{[\code{numeric}]\cr a parameter vector for the null-hypotheses.}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
 \item{\code{statistic}}{the F Test statistic.}
 \item{\code{parameter}}{the degrees of freedom of the appropriate F distribution.}
 \item{\code{p.value}}{the p-value for the test.}
 \item{\code{alternative}}{a character string describing the alternative hypothesis.}
 \item{\code{data.name}}{a character string giving the name of the data.}
 \item{\code{method}}{a character string describing the method.}
}
}
\description{
Performs a test whether a linear model fits to some data via a regression F Test.
}
\details{
Here, a classical F Test for testing the parameter vector of a linear regression
model is done. Only the two sided test is currently supported.
}
\examples{
f.test(y ~ ., data = data.frame(x = rnorm(30), y = rnorm(30)), params = c(1, 1))

}
\seealso{
\code{\link{depth.test}}, \code{\link{sign.test}}
}
