% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiSorting.R
\name{plotMultiSorting}
\alias{plotMultiSorting}
\title{Plot the sorted data (two-dimensional)}
\usage{
plotMultiSorting(
  x,
  xlab,
  ylab,
  main,
  dims = 1:2,
  point.size = 1,
  path.size = 1
)
}
\arguments{
\item{x}{[\code{multiSorting}]\cr An object returned from \code{\link{multiSorting}}.}

\item{xlab}{[\code{character(1)}]\cr Optional x axis label.}

\item{ylab}{[\code{character(1)}]\cr Optional y axis label.}

\item{main}{[\code{character(1)}]\cr Optional title of the plot.}

\item{dims}{[\code{integer(2)}]\cr The dimensions to be plottet. Default is the first two.}

\item{point.size}{[\code{numeric(1)}]\cr Non-negative size of the points in the plot.}

\item{path.size}{[\code{numeric(1)}]\cr Non-negative size of the path in the plot.}
}
\value{
The ggplot object
}
\description{
Plots two dimensions of the data as a scatter plot and the sorting of the
data as lines, which connect the points. Therefore the function ggplot from
the ggplot2-package is used.
}
\examples{
x <- data.frame(x = sort(rnorm(20)), y = sort(rnorm(20)))
multSort <- multiSorting(x)
plotMultiSorting(multSort, point.size = 3, path.size = 2)

}
\seealso{
\code{\link{multiSorting}}
}
