% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signtest.R
\name{sign.test}
\alias{sign.test}
\alias{sign.test.default}
\alias{sign.test.formula}
\title{Sign Test}
\usage{
sign.test(x, ...)

\method{sign.test}{default}(x, ...)

\method{sign.test}{formula}(formula, data, params, ...)
}
\arguments{
\item{x}{[\code{numeric}]\cr vector of residuals.}

\item{...}{[\code{any}]\cr further arguments.}

\item{formula}{[\code{formula}]\cr a formula of the form \code{lhs ~ rhs} where
lhs is a numeric giving the target variable and rhs are the influencing variables .}

\item{data}{[\code{data.frame}]\cr a data frame containing the variables in the formula.}

\item{params}{[\code{numeric}]\cr a parameter vector for the null-hypotheses.}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
 \item{\code{statistic}}{the number of positive residuals.}
 \item{\code{parameter}}{the number of observations.}
 \item{\code{p.value}}{the p-value for the test.}
 \item{\code{alternative}}{a character string describing the alternative hypothesis.}
 \item{\code{data.name}}{a character string giving the name of the data.}
 \item{\code{method}}{a character string describing the method.}
}
}
\description{
Performs a test whether a statistical model fits to some data via a binomial
test on the signs of residuals.
}
\details{
In the default method a residual vector is given and it is tested whether this
residual vector comes from a model that describes the data sufficiently.
With the formula interface it is possible to give a formula and a parameter vector,
so that a linear model with the given parameters is fitted and the residuals are
calculated with this model.

At the moment only linear models for the formula interface are supported!
}
\examples{
sign.test(rnorm(30))
sign.test(y ~ ., data = data.frame(x = rnorm(30), y = rnorm(30)), params = c(1, 1))

}
\seealso{
\code{\link{depth.test}}, \code{\link{f.test}}
}
