% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDepth.R
\name{calcDepth}
\alias{calcDepth}
\title{K Sign Depth}
\usage{
calcDepth(res, K, transform = FALSE, linear = TRUE, exact = TRUE)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{K}{[\code{integer(1)}]\cr
the K of the depth-formula}

\item{transform}{[\code{logical(1)}]\cr
Should the depth be transformed by the formula \eqn{N * (res - 0.5^(K-1))},
so that the distribution of the K Sign Depth converges for large number
of observations N? Default is \code{FALSE}.}

\item{linear}{[code{logical(1)}]\cr
Should the depth be calculated in linear runtime? At the moment, for exact
results this method is available only for K = 2, 3, 4, 5. Default is \code{TRUE}.}

\item{exact}{[code{logical(1)}]\cr
Should the depth be calculated exactly? If \code{FALSE}, terms of scale
o(1) will ne ignored. If \code{linear} is \code{FALSE} always the exact
result is computed, independent of the value of \code{exact}. Default is
\code{TRUE}.}
}
\value{
[\code{numeric(1)}] the calculated K Sign Depth of \code{res}.
}
\description{
\code{calcDepth} calculates the (approximative) K Sign Depth of a given
vector of residuals.
}
\note{
The implementation of the linear calculation is based on a work of
  Kevin Leckey, Dennis Malcherczyk and Christine Mueller.
}
\examples{
calcDepth(rnorm(10), 3)

  # Difference of exact and approximative implementation:
  res <- rnorm(100)
  calcDepth(res, 4, exact = TRUE)
  calcDepth(res, 4, exact = FALSE)
}
