% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDepth.R
\name{calcDepth}
\alias{calcDepth}
\title{Calculating the K-sign depth}
\usage{
calcDepth(res, K, transform = FALSE, linear = TRUE, exact = TRUE)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{K}{[\code{integer(1)}]\cr
the parameter K for the sign depth}

\item{transform}{[\code{logical(1)}]\cr
Should the depth be transformed by the formula \eqn{N * (res - 0.5^(K-1))},
so that the distribution of the K-sign depth converges for large number
of data points N? Default is \code{FALSE}.}

\item{linear}{[\code{logical(1)}]\cr
Should the depth be calculated in linear time complexity? At the moment, for exact
results this method is available only for K = 2, 3, 4, 5. Default is \code{TRUE}.}

\item{exact}{[\code{logical(1)}]\cr
Should the depth be calculated exactly? If \code{FALSE}, terms of scale
o(1) will ne ignored. If \code{linear} is \code{FALSE} always the exact
result is computed, independent of the value of \code{exact}. Default is
\code{TRUE}.}
}
\value{
[\code{numeric(1)}] the calculated K-sign depth of \code{res}.
}
\description{
\code{calcDepth} calculates the (approximative) K-sign depth of a given
vector of residuals.
}
\note{
The implementation of the linear calculation is based on a work of
  Dennis Malcherczyk.
}
\examples{
calcDepth(rnorm(10), 3)

# Difference of exact and approximative implementation:
res <- rnorm(100)
calcDepth(res, 4, exact = TRUE)
calcDepth(res, 4, exact = FALSE)
}
\references{
Horn M. Sign depth for parameter tests in multiple regression.
TU Dortmund University; 2021.

Malcherczy D. K-sign depth: Asymptotic distribution, efficient computation
and applications. TU Dortmund University; 2021.
}
