% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcDepth}
\alias{calcDepth}
\title{Simplicial k-Depth}
\usage{
calcDepth(res, k)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{k}{[\code{integer(1)}]\cr
the k of the depth-formula}
}
\value{
[\code{numeric(1)}] the calculated k-Depth of res.
}
\description{
\code{calcDepth} calculates the k-Depth of a given vector of residuals.
}
\examples{
calcDepth(rnorm(10), 3)
}
