% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDepth.R
\name{calcDepth}
\alias{calcDepth}
\title{Calculating the K-sign depth}
\usage{
calcDepth(res, K, transform = FALSE, linear = TRUE)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{K}{[\code{integer(1)}]\cr
the parameter K for the sign depth}

\item{transform}{[\code{logical(1)}]\cr
Should the depth be transformed by the formula \eqn{N * (res - 0.5^(K-1))},
so that the distribution of the K-sign depth converges for large number
of data points N? Default is \code{FALSE}.}

\item{linear}{[\code{logical(1)}]\cr
Should the depth be calculated in linear time complexity? Default is \code{TRUE}.}
}
\value{
[\code{numeric(1)}] the calculated K-sign depth of \code{res}.
}
\description{
\code{calcDepth} calculates the (approximative) K-sign depth of a given
vector of residuals.
}
\note{
The implementation of the linear calculation is based on a work of
  Dennis Malcherczyk.
}
\examples{
calcDepth(rnorm(10), 3)

# Difference of exact and approximative implementation:
res <- rnorm(100)
calcDepth(res, 4)
calcDepth(res, 4, transform = TRUE)
}
\references{
Horn M. Sign depth for parameter tests in multiple regression.
TU Dortmund University; 2021.

Malcherczy D. K-sign depth: Asymptotic distribution, efficient computation
and applications. TU Dortmund University; 2021.
}
