% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcSimpDepth}
\alias{calcSimpDepth}
\title{Simplified k Sign Depth}
\usage{
calcSimpDepth(res, k)
}
\arguments{
\item{res}{[\code{numeric}]\cr
numeric vector of residuals}

\item{k}{[\code{integer(1)}]\cr
the k of the depth-formula}
}
\value{
[\code{numeric(1)}] the calculated simplified k-Depth of res.
}
\description{
\code{calcSimpDepth} calculates the simplified k Sign Depth of a given vector of residuals.
}
\examples{
calcSimpDepth(rnorm(10), 3)
}
