% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthDistribution.R
\name{depthDistribution}
\alias{depthDistribution}
\alias{qdepth}
\alias{pdepth}
\alias{ddepth}
\title{The Distribution of the Sign Depth}
\usage{
qdepth(p, N, K = 3, transform = FALSE)

pdepth(q, N, K = 3, transform = FALSE)

ddepth(x, N, K = 3, transform = FALSE)
}
\arguments{
\item{p}{[\code{numeric}]\cr vector of probabilities.}

\item{N}{[\code{integer(1)}]\cr number of observations.}

\item{K}{[\code{integer(1)}]\cr parameter of the Sign Depth.
Currently only \eqn{K = 2, 3, 4, 5} is supported. Default is \eqn{K = 3}.}

\item{transform}{[\code{logical(1)}]\cr
Shall the values be transformed so that the distribution converges for large N?}

\item{x, q}{[\code{numeric}]\cr vector of quantiles.}
}
\value{
\code{ddepth} gives the density, \code{pdepth} the distribution function
and \code{qdepth} the quantile function.
}
\description{
Density, distribution function and quantile function for the distribution of the K Sign Depth.
}
\details{
For \eqn{N = K, \ldots, 25} the depth distributions are calculated exactly.
For \eqn{N = 26, \ldots, 100} the distribution is simulated.
For \eqn{N > 100} the values of \eqn{n = 100} are token
(the convergence of the distribution is sufficiently at that point).

If y is a non-transformed value of one of these functions, \eqn{N * (y - 0.5^(K-1))} is the transformed value.
For further understanding of the transformed statistic, see references.
}
\examples{
qdepth(0.05, 20)
ddepth(0.25, 100)
pdepth(0.25, 50)

}
\references{
Kustosz C., Leucht A. and Mueller Ch. H. (2016). Tests based on Simplicial depth for AR(1) models with explosion.
Journal of Time Series Analysis. In press.
}
